+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/output.rb
@@ -108,6 +108,13 @@ class UHT
   Cooling = 'Cooling'
 end
 
+class ULT
+  # Unmet Loads Types
+  HotWaterShowerTime = 'Hot Water Shower Time'
+  HotWaterShowerE = 'Hot Water Shower Energy'
+  HotWaterShowerUnmetTime = 'Hot Water Shower Unmet Time'
+end
+
 class RT
   # Resilience Types
   Battery = 'Battery'