+++ resources/hpxml-measures/HPXMLtoOpenStudio/resources/schedules.rb
@@ -937,6 +937,18 @@ class Schedule
     return '1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0'
   end
 
+  def self.ShowersWeekdayFractions
+    return '0.011,	0.005,	0.003,	0.005,	0.014,	0.052,	0.118,	0.117,	0.095,	0.074,	0.06,	0.047,	0.034,	0.029,	0.026,	0.025,	0.03,	0.039,	0.042,	0.042,	0.042,	0.041,	0.029,	0.021'
+  end
+
+  def self.ShowersWeekendFractions
+    return '0.011,	0.005,	0.003,	0.005,	0.014,	0.052,	0.118,	0.117,	0.095,	0.074,	0.06,	0.047,	0.034,	0.029,	0.026,	0.025,	0.03,	0.039,	0.042,	0.042,	0.042,	0.041,	0.029,	0.021'
+  end
+
+  def self.ShowersMonthlyMultipliers
+    return '1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0'
+  end
+
   def self.GeneralWaterUseWeekdayFractions
     return '0.023, 0.021, 0.021, 0.025, 0.027, 0.038, 0.044, 0.039, 0.037, 0.037, 0.034, 0.035, 0.035, 0.035, 0.039, 0.043, 0.051, 0.064, 0.065, 0.072, 0.073, 0.063, 0.045, 0.034'
   end
@@ -1421,6 +1433,7 @@ class SchedulesFile
     HotWaterDishwasher: Column.new('hot_water_dishwasher', false, true, :frac),
     HotWaterClothesWasher: Column.new('hot_water_clothes_washer', false, true, :frac),
     HotWaterFixtures: Column.new('hot_water_fixtures', true, true, :frac),
+    HotWaterShowers: Column.new('hot_water_showers', true, true, :frac),
     HotWaterRecirculationPump: Column.new('hot_water_recirculation_pump', true, false, :frac),
     GeneralWaterUse: Column.new('general_water_use', true, false, :frac),
     Sleeping: Column.new('sleeping', false, false, nil),