+++ resources/hpxml-measures/ReportSimulationOutput/measure.rb
@@ -97,6 +97,12 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     arg.setDefaultValue(true)
     args << arg
 
+    arg = OpenStudio::Measure::OSArgument::makeBoolArgument('include_annual_unmet_loads', false)
+    arg.setDisplayName('Generate Annual Output: Unmet Loads')
+    arg.setDescription('Generates annual unmet loads for hot water.')
+    arg.setDefaultValue(true)
+    args << arg
+
     arg = OpenStudio::Measure::OSArgument::makeBoolArgument('include_annual_peak_fuels', false)
     arg.setDisplayName('Generate Annual Output: Peak Fuels')
     arg.setDescription('Generates annual electricity peaks for summer/winter.')
@@ -290,6 +296,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
      @end_uses,
      @loads,
      @unmet_hours,
+     @unmet_loads,
      @peak_fuels,
      @peak_loads,
      @component_loads,
@@ -348,6 +355,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     end
 
     unmet_hours_program = @model.getEnergyManagementSystemPrograms.find { |p| p.additionalProperties.getFeatureAsString('ObjectType').to_s == Constants.ObjectNameUnmetHoursProgram }
+    unmet_loads_program = @model.getModelObjectByName(Constants.ObjectNameUnmetLoadsProgram.gsub(' ', '_')).get.to_EnergyManagementSystemProgram.get
     total_loads_program = @model.getEnergyManagementSystemPrograms.find { |p| p.additionalProperties.getFeatureAsString('ObjectType').to_s == Constants.ObjectNameTotalLoadsProgram }
     comp_loads_program = @model.getEnergyManagementSystemPrograms.find { |p| p.additionalProperties.getFeatureAsString('ObjectType').to_s == Constants.ObjectNameComponentLoadsProgram }
     heated_zones = eval(@model.getBuilding.additionalProperties.getFeatureAsString('heated_zones').get)
@@ -452,6 +460,20 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       if args[:include_timeseries_unmet_hours]
         result << OpenStudio::IdfObject.load("EnergyManagementSystem:OutputVariable,#{unmet_hour.ems_variable}_timeseries_outvar,#{unmet_hour.ems_variable},Summed,ZoneTimestep,#{unmet_hours_program.name},hr;").get
         result << OpenStudio::IdfObject.load("Output:Variable,*,#{unmet_hour.ems_variable}_timeseries_outvar,#{args[:timeseries_frequency]};").get
+      end
+    end
+
+    # Unmet Loads (annual only)
+    @unmet_loads.each do |key, unmet_load|
+      units = 'hr'
+      if key == ULT::HotWaterShowerE
+        units = 'J'
+      end
+      result << OpenStudio::IdfObject.load("EnergyManagementSystem:OutputVariable,#{unmet_load.ems_variable}_annual_outvar,#{unmet_load.ems_variable},Summed,SystemTimestep,#{unmet_loads_program.name},#{units};").get
+      result << OpenStudio::IdfObject.load("Output:Variable,*,#{unmet_load.ems_variable}_annual_outvar,runperiod;").get
+      if args[:include_timeseries_unmet_loads]
+        result << OpenStudio::IdfObject.load("EnergyManagementSystem:OutputVariable,#{unmet_load.ems_variable}_timeseries_outvar,#{unmet_load.ems_variable},Summed,SystemTimestep,#{unmet_loads_program.name},hr;").get
+        result << OpenStudio::IdfObject.load("Output:Variable,*,#{unmet_load.ems_variable}_timeseries_outvar,#{args[:timeseries_frequency]};").get
       end
     end
 
@@ -783,6 +805,14 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       unmet_hour.annual_output = get_report_variable_data_annual(['EMS'], ["#{unmet_hour.ems_variable}_annual_outvar"], 1.0)
       if args[:include_timeseries_unmet_hours]
         unmet_hour.timeseries_output = get_report_variable_data_timeseries(['EMS'], ["#{unmet_hour.ems_variable}_timeseries_outvar"], 1.0, 0, args[:timeseries_frequency])
+      end
+    end
+
+    # Unmet Loads
+    @unmet_loads.each do |_key, unmet_load|
+      unmet_load.annual_output = get_report_variable_data_annual(['EMS'], ["#{unmet_load.ems_variable}_annual_outvar"], 1.0)
+      if args[:include_timeseries_unmet_loads]
+        unmet_load.timeseries_output = get_report_variable_data_timeseries(['EMS'], ["#{unmet_load.ems_variable}_timeseries_outvar"], 1.0, 0, args[:timeseries_frequency])
       end
     end
 
@@ -1542,6 +1572,14 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       results_out << [line_break]
     end
 
+    # Unmet loads
+    if args[:include_annual_unmet_loads]
+      @unmet_loads.each do |_load_type, unmet_load|
+        results_out << ["#{unmet_load.name} (#{unmet_load.annual_units})", unmet_load.annual_output.to_f.round(n_digits)]
+      end
+      results_out << [line_break]
+    end
+
     # Peak fuels
     if args[:include_annual_peak_fuels]
       @peak_fuels.each do |_key, peak_fuel|
@@ -1796,6 +1834,11 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     else
       unmet_hours_data = []
     end
+    if args[:include_timeseries_unmet_loads]
+      unmet_loads_data = @unmet_loads.values.map { |x| [x.name, x.timeseries_units] + x.timeseries_output.map { |v| v.round(n_digits) } }
+    else
+      unmet_loads_data = []
+    end
     if args[:include_timeseries_zone_temperatures]
       zone_temps_data = @zone_temps.values.select { |x| x.timeseries_output.sum(0.0) != 0 }.map { |x| [x.name, x.timeseries_units] + x.timeseries_output.map { |v| v.round(n_digits) } }
     else
@@ -1825,7 +1868,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     end
 
     return if (total_energy_data.size + fuel_data.size + end_use_data.size + system_use_data.size + emissions_data.size + emission_fuel_data.size +
-               emission_end_use_data.size + hot_water_use_data.size + total_loads_data.size + comp_loads_data.size + unmet_hours_data.size +
+               emission_end_use_data.size + hot_water_use_data.size + total_loads_data.size + comp_loads_data.size + unmet_hours_data.size + unmet_loads_data.size +
                zone_temps_data.size + airflows_data.size + weather_data.size + resilience_data.size + output_variables_data.size) == 0
 
     fail 'Unable to obtain timestamps.' if @timestamps.empty?
@@ -1834,7 +1877,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       # Assemble data
       data = data.zip(*timestamps2, *timestamps3, *total_energy_data, *fuel_data, *end_use_data, *system_use_data, *emissions_data,
                       *emission_fuel_data, *emission_end_use_data, *hot_water_use_data, *total_loads_data, *comp_loads_data,
-                      *unmet_hours_data, *zone_temps_data, *airflows_data, *weather_data, *resilience_data, *output_variables_data)
+                      *unmet_hours_data, *unmet_loads_data, *zone_temps_data, *airflows_data, *weather_data, *resilience_data, *output_variables_data)
 
       # Error-check
       n_elements = []
@@ -1897,7 +1940,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       h['TimeUTC'] = timestamps3[2..-1] if timestamps_utc
 
       [total_energy_data, fuel_data, end_use_data, system_use_data, emissions_data, emission_fuel_data,
-       emission_end_use_data, hot_water_use_data, total_loads_data, comp_loads_data, unmet_hours_data,
+       emission_end_use_data, hot_water_use_data, total_loads_data, comp_loads_data, unmet_hours_data, unmet_loads_data,
        zone_temps_data, airflows_data, weather_data, resilience_data, output_variables_data].each do |d|
         d.each do |o|
           grp, name = o[0].split(':', 2)
@@ -2268,6 +2311,14 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     attr_accessor(:ems_variable)
   end
 
+  class UnmetLoads < BaseOutput
+    def initialize(ems_variable:)
+      super()
+      @ems_variable = ems_variable
+    end
+    attr_accessor(:ems_variable)
+  end
+
   class IdealLoad < BaseOutput
     def initialize(variables: [])
       super()
@@ -2603,6 +2654,22 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       unmet_hour.name = "Unmet Hours: #{load_type}"
       unmet_hour.annual_units = 'hr'
       unmet_hour.timeseries_units = 'hr'
+    end
+
+    # Unmet Loads
+    @unmet_loads = {}
+    @unmet_loads[ULT::HotWaterShowerTime] = UnmetLoads.new(ems_variable: 'ShowerTime')
+    @unmet_loads[ULT::HotWaterShowerE] = UnmetLoads.new(ems_variable: 'ShowerE')
+    @unmet_loads[ULT::HotWaterShowerUnmetTime] = UnmetLoads.new(ems_variable: 'ShowerSagTime')
+
+    @unmet_loads.each do |load_type, unmet_load|
+      units = 'hr'
+      if load_type == ULT::HotWaterShowerE
+        units = 'J'
+      end
+      unmet_load.name = "Unmet Loads: #{load_type}"
+      unmet_load.annual_units = units
+      unmet_load.timeseries_units = units
     end
 
     # Peak Loads